feat: session rotate

+++ src/Services/SessionService.php
@@ -2,6 +2,7 @@
 
 namespace Secra\Services;
 
+use Exception;
 use Secra\Arch\DI\Attributes\Inject;
 use Secra\Arch\DI\Attributes\Provide;
 use Secra\Arch\DI\Attributes\Singleton;
@@ -25,13 +26,44 @@ class SessionService
   )
   {
     session_start();
-    $session_id = $_COOKIE['session_id'] ?? $_SESSION['session_id'] ?? null;
-    if ($session_id) {
-      $this->currentSession = $this->validateSession($session_id);
-      if ($this->currentSession) {
+    if (isset($_SESSION['session_id'])) {
+      // There is a active session, just validate it
+      if ($this->currentSession = $this->validateSession($_SESSION['session_id'])) {
         $this->currentUser = $this->userRepository->getUserById($this->currentSession->user_id);
       }
+    } elseif (isset($_COOKIE['session_id'])) {
+      // There is a session id in the cookie, validate and rotate it
+      if ($this->currentSession = $this->validateSession($_COOKIE['session_id'])) {
+        $this->currentUser = $this->userRepository->getUserById($this->currentSession->user_id);
+        $this->rotateCurrentSession();
+      }
     }
+  }
+
+  /**
+   * Rotates the current session and changes the session id to enhance security
+   */
+  private function rotateCurrentSession(): void
+  {
+    if (!$this->getCurrentSession()) {
+      return;
+    }
+    $current_session_id = $this->getCurrentSession()->session_id;
+    $this->sessionRepository->delete($current_session_id);
+
+    $new_session_id = session_create_id();
+    $new_session = new Session();
+    $new_session->session_id = $new_session_id;
+    $new_session->user_id = $this->getCurrentSession()->user_id;
+    $new_session->expires_at = $this->getCurrentSession()->expires_at;
+
+    $this->currentSession = $this->sessionRepository->save($new_session);
+    $expires_at = strtotime($this->getCurrentSession()->expires_at);
+
+    setcookie('session_id', $new_session_id, $expires_at, '/', '', false, true);
+    $_SESSION['session_id'] = $new_session_id;
+
+    $this->logger->info('Rotated session id from ' . $current_session_id . ' to ' . $new_session_id);
   }
 
   public function isUserLoggedIn(): bool
@@ -70,7 +102,7 @@ class SessionService
     $current_session_id = $_COOKIE['session_id'] ?? $_SESSION['session_id'] ?? null;
     if ($current_session_id) {
       $oldSession = $this->sessionRepository->getSessionById($current_session_id);
-      if ($oldSession) {
+      if ($oldSession && $oldSession->user_id === $user->user_id) {
         $this->sessionRepository->delete($current_session_id);
       }
     }
@@ -83,11 +115,15 @@ class SessionService
     $session->session_id = $session_id;
     $session->user_id = $user->user_id;
     $session->expires_at = date('Y-m-d H:i:s', strtotime('+30 day'));
-    $this->sessionRepository->save($session);
+    $session = $this->sessionRepository->save($session);
+
+    if (!$session) {
+      throw new Exception('Failed to create session');
+    }
 
     setcookie('session_id', $session->session_id, strtotime('+30 day'), '/', '', false, true);
 
-    return $this->sessionRepository->getSessionById($session->session_id);
+    return $session;
   }
 
   public function destroyCurrentSession(): void
@@ -98,5 +134,7 @@ class SessionService
     }
     setcookie('session_id', '', time() - 3600, '/', '', false, true);
     session_destroy();
+    $this->currentSession = null;
+    $this->currentUser = null;
   }
 }
